+++ contracts/KYCFactory.sol
@@ -8,12 +8,12 @@ import "./KYC2.sol";
 
 library KYCFactory {
 
-        //event createkycF(string    id,address storageAddress,uint16 company_registry_id ,uint timestamp);
+        //event createkycF(string    id,address storageAddress,uint64 company_registry_id ,uint timestamp);
 
 
 
 
-    function createKyc( string memory id,uint16 _kycPerformer_registry_id, string[] memory   attributesList, string[] memory   attributesVal) public returns(address)
+    function createKyc( string memory id,uint64 _kycPerformer_registry_id, string[] memory   attributesList, string[] memory   attributesVal) public returns(address)
     {
                 KYCInt  kyc = new KYC( address(this) , id,_kycPerformer_registry_id,attributesList,    attributesVal);
                 address kycaddr= address(kyc);
@@ -24,7 +24,7 @@ library KYCFactory {
 
 
 /*
-    function createKyc2( string memory id,uint16 _kycPerformer_registry_id, string[] memory   attributesList, string[] memory   attributesVal) public returns(address)
+    function createKyc2( string memory id,uint64 _kycPerformer_registry_id, string[] memory   attributesList, string[] memory   attributesVal) public returns(address)
     {
                 KYCInt  kyc = new KYC2( address(this) , id,_kycPerformer_registry_id,attributesList,    attributesVal);
                 address kycaddr= address(kyc);
