+++ contracts/KYCLib.sol
@@ -10,7 +10,7 @@ library KYCLib {
 
 
 
-               function init(address _KYCStorage  ,string memory id, uint16 _kycIssuer,string[] memory   attributesList, string[] memory attributesVal)   public
+               function init(address _KYCStorage  ,string memory id, uint64 _kycIssuer,string[] memory   attributesList, string[] memory attributesVal)   public
                    {
 
                    KYCStorage(_KYCStorage).setKYCIssuer(_kycIssuer);
@@ -29,7 +29,7 @@ library KYCLib {
                    }
 
 
-           function setAttributePermission(address _KYCStorage  ,string memory  attributeName ,uint16 companion_id , uint8 permission)  public
+           function setAttributePermission(address _KYCStorage  ,string memory  attributeName ,uint64 companion_id , uint8 permission)  public
                {
 
                    KYCStorage(_KYCStorage).setAttributePermission(attributeName , companion_id ,  permission);
@@ -40,7 +40,7 @@ library KYCLib {
 
 
 
-                   function getKYCIssuer(address _KYCStorage  )  view   external  returns (uint16)
+                   function getKYCIssuer(address _KYCStorage  )  view   external  returns (uint64)
                    {
                        return KYCStorage(_KYCStorage).getKYCIssuer();
                    }
@@ -62,7 +62,7 @@ library KYCLib {
 
 
 
-                   function isAttributePermited(address _KYCStorage  ,string memory  attrName,uint16 companion_id) view public  returns (uint8)
+                   function isAttributePermited(address _KYCStorage  ,string memory  attrName,uint64 companion_id) view public  returns (uint8)
                    {
                            return KYCStorage(_KYCStorage).isAttributePermited(attrName, companion_id);
 
@@ -70,7 +70,7 @@ library KYCLib {
 
 
 
-                   function getAttribute(address _KYCStorage ,string memory attrName,uint16 companion_id) view public  returns (bytes32 )
+                   function getAttribute(address _KYCStorage ,string memory attrName,uint64 companion_id) view public  returns (bytes32 )
                    {
 
                        return KYCStorage(_KYCStorage).getAttribute(attrName, companion_id);
